third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/source-frame-count-expected.txt

Tests that scripts panel does not create too many source frames.


Running: testSourceFramesCount
Page reloaded.
